Offerte printen

+++ src/MilieuKorting.java
@@ -1,5 +1,5 @@
 public class MilieuKorting {
-    protected int kortingPercentage;
+    protected double kortingPercentage;
     protected String beschrijving;
 
     public MilieuKorting(int kortingPercentage) {
@@ -10,7 +10,7 @@ public class MilieuKorting {
         this.beschrijving = beschrijving;
     }
 
-    public int getKortingsPercentage() {return kortingPercentage;}
+    public double getKortingsPercentage() {return kortingPercentage;}
     public void setKortingsPercentage(int kortingsPercentage) {this.kortingPercentage = kortingsPercentage;}
     public String getBeschrijving() {return beschrijving;}
     public void setBeschrijving(String beschrijving) {this.beschrijving = beschrijving;}
